[Silver III] Title: 주유소, Time: 332 ms, Memory: 42076 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -8,19 +8,19 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         int N = Integer.parseInt(br.readLine());
         StringTokenizer st = new StringTokenizer(br.readLine());
-        int[] distance = new int[N-1];
-        int[] price = new int[N];
+        long[] distance = new long[N-1];
+        long[] price = new long[N];
         for (int i = 0; i < N-1; i++) {
-            distance[i] = Integer.parseInt(st.nextToken());
+            distance[i] = Long.parseLong(st.nextToken());
         }
 
         st = new StringTokenizer(br.readLine());
         for (int i = 0; i < N; i++) {
-            price[i] = Integer.parseInt(st.nextToken());
+            price[i] = Long.parseLong(st.nextToken());
         }
 
-        int total = distance[0] * price[0];
-        int minPrice = price[0];
+        long total = distance[0] * price[0];
+        long minPrice = price[0];
         for (int i = 1; i < N-1; i++) {
             if (price[i] < minPrice) {
                 minPrice = price[i];
